Use our own popcorn-wrapper module, media-popcorn-wrapper.js.
the SEAMLESS X-TAG MODALITY and butter-bar.js already do a lot of what
popcorn-wrapper did in terms of creating media, so media-popcorn-wrapper
is designed to take a pre-existing media element, wrap it in a Popcorn
instance, and hand it over to butter.

This relieves us from having to manually delegate events from the
preview frame's media to its "mirror" in butter, as they're now one
and the same. This means that butter's scrubber is precisely synchronized
with the media element in the preview pane.