LivePreview now exposes editor and preview channels.
This will make it a lot easier to support UIs that bridge between
the preview frame and editor once we move the preview frame into a
separate origin. See the changes to preview-to-editor-mapping.js
for an example of how this can be done to support both
same-origin and cross-origin preview frames.