Added a 'preview plugin' architecture.
This makes it easier to support functionality that enhances the preview
area but also needs to talk to the editor, and which may or may not be
running in an environment where the preview frame is in a cross-origin
document.

The only preview plugin right now is the 'source-higlighter', aka
x-ray goggles effect.