Added 'previewFrameURL' option to friendlycode and stub preview proxy module.
After talking to stenington about this, we realized that it'd be a lot easier
to understand how the live preview proxying situation worked if we actually
made a stub implementation that the example filled-out with a proof-of-concept,
non-production implementation. This commit adds a 'previewFrameURL' option
to friendlycode, which by default causes the stub implementation to be
loaded and an exception to be thrown. However, the example replaces the
stub with a functional proxy that does real cross-origin communication with
a preview frame.

Also added some comments here and there to hopefully makes this all
less confusing.